backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
SQLAlchemy>=2.0
psycopg2-binary
alembic
boto3
pyjwt
passlib[bcrypt]
